+++ trustm_installation_script.sh
@@ -1,13 +1,25 @@
 #!/bin/bash
 
-# SPDX-FileCopyrightText: Copyright (c) 2025 Infineon Technologies AG
+# SPDX-FileCopyrightText: Copyright (c) 2026 Infineon Technologies AG
 # SPDX-License-Identifier: MIT
 
 sudo apt update 
-sudo apt -y install git gcc libssl-dev gpiod libgpiod-dev curl
+sudo apt -y install git gcc libssl-dev gpiod libgpiod-dev curl \
+                    python3 python3-jinja2 python3-jsonschema perl
+
+(
+    echo "-----> Generate mbedtls config files" 
+    cd external/optiga-trust-m/external/mbedtls-4.x 
+    python3 framework/scripts/make_generated_files.py
+)
+(
+    echo "-----> Generate PSA config files"
+    cd external/optiga-trust-m/external/mbedtls-4.x/tf-psa-crypto
+    python3 framework/scripts/make_generated_files.py
+)
 
 set -e
-echo "-----> Build Trust M Linux Tools and provider"
+echo "-----> Build Trust M provider"
 sudo make uninstall
 make clean
 make -j5
@@ -20,7 +32,7 @@ export IFX_CERT_KEY=certificates/OPTIGA_Trust_M_Infineon_Test_CA_Key.pem
 
 declare -i ErrorCount=0
 
-clear
+clear 2>/dev/null || true
 
 if [ -e $IFX_CERT_KEY ]
 then