+++ CHANGELOG.md
@@ -10,6 +10,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.4] - 2026-08-07
+
+### Changed
+
+- Updated optiga-trust-m library to version 5.8.0
+- Migrate mbedtls support from mbedtls2.x/3.x to mbedtls 4.x
+
 ## [1.0.3] - 2026-05-26
 
 ### Changed